requirements_okx.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anyio==4.9.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
Automat==24.8.1
backports.tarfile==1.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
Cython==3.0.12
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
loguru==0.7.3
MarkupSafe==3.0.2
more-itertools==10.6.0
multidict==6.2.0
packaging==24.2
pg8000==1.31.2
propcache==0.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-okx==0.3.5
python-socketio==5.12.1
redis==5.2.1
requests==2.32.3
scramp==1.4.5
SecretStorage==3.3.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2